+++ blockchain/txlookup.go
@@ -199,7 +199,7 @@ func disconnectTransactions(txStore TxStore, block *dcrutil.Block, parent *dcrut
 		// sides of a fork would otherwise not be updated.
 		if txD, exists := txStore[*tx.Sha()]; exists {
 			txD.Tx = nil
-			txD.BlockHeight = int64(wire.NullBlockHeight)
+			txD.BlockHeight = int32(wire.NullBlockHeight)
 			txD.BlockIndex = wire.NullBlockIndex
 			txD.Spent = nil
 			txD.Err = database.ErrTxShaMissing
@@ -244,7 +244,7 @@ func disconnectTransactions(txStore TxStore, block *dcrutil.Block, parent *dcrut
 				// sides of a fork would otherwise not be updated.
 				if txD, exists := txStore[*tx.Sha()]; exists {
 					txD.Tx = nil
-					txD.BlockHeight = int64(wire.NullBlockHeight)
+					txD.BlockHeight = int32(wire.NullBlockHeight)
 					txD.BlockIndex = wire.NullBlockIndex
 					txD.Spent = nil
 					txD.Err = database.ErrTxShaMissing
